requirements.txt
Pillow
pillow-heif
pyexiftool
scipy